Reorganize the files
